third_party/WebKit/LayoutTests/external/wpt/web-animations/animation-model/animation-types/accumulation-per-property-expected.txt

This is a testharness.js-based test.
Found 566 tests; 396 PASS, 170 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Setup
PASS align-content (type: discrete) has testAccumulation function
PASS align-content: "flex-end" onto "flex-start"
PASS align-content: "flex-start" onto "flex-end"
PASS align-items (type: discrete) has testAccumulation function
PASS align-items: "flex-end" onto "flex-start"
PASS align-items: "flex-start" onto "flex-end"
PASS align-self (type: discrete) has testAccumulation function
PASS align-self: "flex-end" onto "flex-start"
PASS align-self: "flex-start" onto "flex-end"
PASS backface-visibility (type: discrete) has testAccumulation function
PASS backface-visibility: "hidden" onto "visible"
PASS backface-visibility: "visible" onto "hidden"
PASS background-attachment (type: discrete) has testAccumulation function
PASS background-attachment: "local" onto "fixed"
PASS background-attachment: "fixed" onto "local"
PASS background-color (type: color) has testAccumulation function
FAIL background-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL background-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL background-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL background-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL background-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL background-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS background-blend-mode (type: discrete) has testAccumulation function
PASS background-blend-mode: "screen" onto "multiply"
PASS background-blend-mode: "multiply" onto "screen"
PASS background-clip (type: discrete) has testAccumulation function
PASS background-clip: "content-box" onto "padding-box"
PASS background-clip: "padding-box" onto "content-box"
PASS background-image (type: discrete) has testAccumulation function
PASS background-image: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS background-image: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS background-origin (type: discrete) has testAccumulation function
PASS background-origin: "content-box" onto "padding-box"
PASS background-origin: "padding-box" onto "content-box"
PASS background-repeat (type: discrete) has testAccumulation function
PASS background-repeat: "round" onto "space"
PASS background-repeat: "space" onto "round"
PASS border-bottom-color (type: color) has testAccumulation function
FAIL border-bottom-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-bottom-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-bottom-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-bottom-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-bottom-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-bottom-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS border-bottom-style (type: discrete) has testAccumulation function
PASS border-bottom-style: "solid" onto "dotted"
PASS border-bottom-style: "dotted" onto "solid"
PASS border-bottom-width (type: length) has testAccumulation function
FAIL border-bottom-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL border-bottom-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS border-collapse (type: discrete) has testAccumulation function
PASS border-collapse: "separate" onto "collapse"
PASS border-collapse: "collapse" onto "separate"
PASS border-image-outset (type: discrete) has testAccumulation function
PASS border-image-outset: "5 6 7 8" onto "1 2 3 4"
PASS border-image-outset: "1 2 3 4" onto "5 6 7 8"
PASS border-image-repeat (type: discrete) has testAccumulation function
FAIL border-image-repeat: "repeat repeat" onto "stretch stretch" assert_equals: The value should be repeat repeat at 0ms expected "repeat repeat" but got "repeat"
FAIL border-image-repeat: "stretch stretch" onto "repeat repeat" assert_equals: The value should be stretch stretch at 0ms expected "stretch stretch" but got "stretch"
PASS border-image-slice (type: discrete) has testAccumulation function
PASS border-image-slice: "5 6 7 8" onto "1 2 3 4"
PASS border-image-slice: "1 2 3 4" onto "5 6 7 8"
PASS border-image-source (type: discrete) has testAccumulation function
PASS border-image-source: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS border-image-source: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS border-image-width (type: discrete) has testAccumulation function
PASS border-image-width: "5 6 7 8" onto "1 2 3 4"
PASS border-image-width: "1 2 3 4" onto "5 6 7 8"
PASS border-left-color (type: color) has testAccumulation function
FAIL border-left-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-left-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-left-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-left-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-left-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-left-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS border-left-style (type: discrete) has testAccumulation function
PASS border-left-style: "solid" onto "dotted"
PASS border-left-style: "dotted" onto "solid"
PASS border-left-width (type: length) has testAccumulation function
FAIL border-left-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL border-left-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS border-right-color (type: color) has testAccumulation function
FAIL border-right-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-right-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-right-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-right-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-right-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-right-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS border-right-style (type: discrete) has testAccumulation function
PASS border-right-style: "solid" onto "dotted"
PASS border-right-style: "dotted" onto "solid"
PASS border-right-width (type: length) has testAccumulation function
FAIL border-right-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL border-right-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS border-spacing (type: lengthPair) has testAccumulation function
FAIL border-spacing: length pair assert_equals: The value should be 20px 20px at 0ms expected "20px 20px" but got "10px 10px"
FAIL border-spacing: length pair of rem assert_equals: The value should be 20px 20px at 0ms expected "20px 20px" but got "10px 10px"
PASS border-top-color (type: color) has testAccumulation function
FAIL border-top-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-top-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-top-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL border-top-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-top-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL border-top-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS border-top-style (type: discrete) has testAccumulation function
PASS border-top-style: "solid" onto "dotted"
PASS border-top-style: "dotted" onto "solid"
PASS border-top-width (type: length) has testAccumulation function
FAIL border-top-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL border-top-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS box-shadow (type: boxShadowList) has testAccumulation function
FAIL box-shadow: shadow assert_equals: The value should be rgb(240, 240, 240) 20px 20px 20px 20px at 0ms expected "rgb(240, 240, 240) 20px 20px 20px 20px" but got "rgb(120, 120, 120) 10px 10px 10px 10px"
PASS box-sizing (type: discrete) has testAccumulation function
PASS box-sizing: "border-box" onto "content-box"
PASS box-sizing: "content-box" onto "border-box"
PASS caption-side (type: discrete) has testAccumulation function
PASS caption-side: "bottom" onto "top"
PASS caption-side: "top" onto "bottom"
PASS caret-color (type: color) has testAccumulation function
FAIL caret-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL caret-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL caret-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL caret-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL caret-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL caret-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS clear (type: discrete) has testAccumulation function
PASS clear: "right" onto "left"
PASS clear: "left" onto "right"
PASS clip (type: rect) has testAccumulation function
FAIL clip: rect assert_equals: The value should be rect(110px, 110px, 110px, 110px) at 0ms expected "rect(110px, 110px, 110px, 110px)" but got "rect(10px, 10px, 10px, 10px)"
PASS clip (type: discrete) has testAccumulation function
PASS clip: "auto" onto "rect(10px, 10px, 10px, 10px)"
PASS clip: "rect(10px, 10px, 10px, 10px)" onto "auto"
PASS clip: "rect(10px, 10px, 10px, auto)" onto "rect(10px, 10px, 10px, 10px)"
PASS clip: "rect(10px, 10px, 10px, 10px)" onto "rect(10px, 10px, 10px, auto)"
PASS clip-rule (type: discrete) has testAccumulation function
PASS clip-rule: "nonzero" onto "evenodd"
PASS clip-rule: "evenodd" onto "nonzero"
PASS color (type: color) has testAccumulation function
FAIL color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS color-interpolation (type: discrete) has testAccumulation function
PASS color-interpolation: "auto" onto "linearrgb"
PASS color-interpolation: "linearrgb" onto "auto"
PASS color-interpolation-filters (type: discrete) has testAccumulation function
PASS color-interpolation-filters: "linearrgb" onto "srgb"
PASS color-interpolation-filters: "srgb" onto "linearrgb"
PASS column-count (type: positiveInteger) has testAccumulation function
FAIL column-count: positive integer assert_equals: The value should be 3 at 0ms expected "3" but got "2"
PASS column-count (type: discrete) has testAccumulation function
PASS column-count: "10" onto "auto"
PASS column-count: "auto" onto "10"
PASS column-gap (type: length) has testAccumulation function
FAIL column-gap: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL column-gap: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS column-gap (type: discrete) has testAccumulation function
PASS column-gap: "200px" onto "normal"
PASS column-gap: "normal" onto "200px"
PASS column-rule-color (type: color) has testAccumulation function
FAIL column-rule-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL column-rule-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL column-rule-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL column-rule-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL column-rule-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL column-rule-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS column-fill (type: discrete) has testAccumulation function
PASS column-fill: "balance" onto "auto"
PASS column-fill: "auto" onto "balance"
PASS column-rule-style (type: discrete) has testAccumulation function
PASS column-rule-style: "dotted" onto "none"
PASS column-rule-style: "none" onto "dotted"
PASS column-rule-width (type: length) has testAccumulation function
FAIL column-rule-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL column-rule-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS column-width (type: length) has testAccumulation function
FAIL column-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL column-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS column-width (type: discrete) has testAccumulation function
PASS column-width: "1px" onto "auto"
PASS column-width: "auto" onto "1px"
PASS content (type: discrete) has testAccumulation function
PASS content: ""b"" onto ""a""
PASS content: ""a"" onto ""b""
PASS counter-increment (type: discrete) has testAccumulation function
PASS counter-increment: "ident-2 2" onto "ident-1 1"
PASS counter-increment: "ident-1 1" onto "ident-2 2"
PASS counter-reset (type: discrete) has testAccumulation function
PASS counter-reset: "ident-2 2" onto "ident-1 1"
PASS counter-reset: "ident-1 1" onto "ident-2 2"
PASS cursor (type: discrete) has testAccumulation function
PASS cursor: "wait" onto "pointer"
PASS cursor: "pointer" onto "wait"
PASS dominant-baseline (type: discrete) has testAccumulation function
PASS dominant-baseline: "alphabetic" onto "ideographic"
PASS dominant-baseline: "ideographic" onto "alphabetic"
PASS empty-cells (type: discrete) has testAccumulation function
PASS empty-cells: "hide" onto "show"
PASS empty-cells: "show" onto "hide"
PASS fill-opacity (type: opacity) has testAccumulation function
FAIL fill-opacity: [0, 1] number assert_equals: The value should be 0.6 at 0ms expected "0.6" but got "0.3"
FAIL fill-opacity: [0, 1] number (clamped) assert_equals: The value should be 1 at 0ms expected "1" but got "0.3"
PASS fill-rule (type: discrete) has testAccumulation function
PASS fill-rule: "nonzero" onto "evenodd"
PASS fill-rule: "evenodd" onto "nonzero"
PASS filter (type: filterList) has testAccumulation function
FAIL filter: same ordered filter functions assert_equals: The value should be blur(30px) brightness(0) at 0ms expected "blur(30px) brightness(0)" but got "blur(20px) brightness(0.1)"
PASS filter: mismatched ordered filter functions
PASS flex-basis (type: lengthPercentageOrCalc) has testAccumulation function
PASS flex-basis: length
PASS flex-basis: length of rem
PASS flex-basis: percentage
FAIL flex-basis: units "%" onto "px" assert_equals: The value should be calc(10px + 10%) at 0ms expected "calc(10px + 10%)" but got "10%"
FAIL flex-basis: units "px" onto "%" assert_equals: The value should be calc(10px + 10%) at 0ms expected "calc(10px + 10%)" but got "10px"
FAIL flex-basis: units "rem" onto "%" assert_equals: The value should be calc(20px + 10%) at 0ms expected "calc(20px + 10%)" but got "20px"
FAIL flex-basis: units "%" onto "rem" assert_equals: The value should be calc(20px + 10%) at 0ms expected "calc(20px + 10%)" but got "10%"
FAIL flex-basis: units "rem" onto "em" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL flex-basis: units "em" onto "rem" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL flex-basis: units "calc" onto "px" assert_equals: The value should be calc(30px + 20%) at 0ms expected "calc(30px + 20%)" but got "calc(20px + 20%)"
FAIL flex-basis: calc assert_equals: The value should be calc(30px + 30%) at 0ms expected "calc(30px + 30%)" but got "calc(20px + 20%)"
PASS flex-basis (type: discrete) has testAccumulation function
PASS flex-basis: "10px" onto "auto"
PASS flex-basis: "auto" onto "10px"
PASS flex-direction (type: discrete) has testAccumulation function
PASS flex-direction: "row-reverse" onto "row"
PASS flex-direction: "row" onto "row-reverse"
PASS flex-grow (type: positiveNumber) has testAccumulation function
FAIL flex-grow: positive number assert_equals: The value should be 2.2 at 0ms expected "2.2" but got "1.1"
PASS flex-shrink (type: positiveNumber) has testAccumulation function
FAIL flex-shrink: positive number assert_equals: The value should be 2.2 at 0ms expected "2.2" but got "1.1"
PASS flex-wrap (type: discrete) has testAccumulation function
PASS flex-wrap: "wrap" onto "nowrap"
PASS flex-wrap: "nowrap" onto "wrap"
PASS flood-color (type: color) has testAccumulation function
FAIL flood-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL flood-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL flood-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL flood-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL flood-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL flood-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS flood-opacity (type: opacity) has testAccumulation function
FAIL flood-opacity: [0, 1] number assert_equals: The value should be 0.6 at 0ms expected "0.6" but got "0.3"
FAIL flood-opacity: [0, 1] number (clamped) assert_equals: The value should be 1 at 0ms expected "1" but got "0.3"
PASS font-stretch (type: percentage) has testAccumulation function
FAIL font-stretch: percentage assert_equals: The value should be 130% at 0ms expected "130%" but got "70%"
PASS font-style (type: discrete) has testAccumulation function
FAIL font-style: "oblique" onto "italic" assert_equals: The value should be oblique at 0ms expected "oblique" but got "italic"
PASS font-style: "italic" onto "oblique"
PASS float (type: discrete) has testAccumulation function
PASS float: "right" onto "left"
PASS float: "left" onto "right"
PASS font-family (type: discrete) has testAccumulation function
PASS font-family: "verdana" onto "helvetica"
PASS font-family: "helvetica" onto "verdana"
PASS font-feature-settings (type: discrete) has testAccumulation function
PASS font-feature-settings: "normal" onto ""liga" 5"
PASS font-feature-settings: ""liga" 5" onto "normal"
PASS font-kerning (type: discrete) has testAccumulation function
PASS font-kerning: "normal" onto "auto"
PASS font-kerning: "auto" onto "normal"
PASS font-variant-caps (type: discrete) has testAccumulation function
PASS font-variant-caps: "unicase" onto "small-caps"
PASS font-variant-caps: "small-caps" onto "unicase"
PASS font-variant-east-asian (type: discrete) has testAccumulation function
PASS font-variant-east-asian: "proportional-width" onto "full-width"
PASS font-variant-east-asian: "full-width" onto "proportional-width"
PASS font-variant-ligatures (type: discrete) has testAccumulation function
PASS font-variant-ligatures: "no-common-ligatures" onto "common-ligatures"
PASS font-variant-ligatures: "common-ligatures" onto "no-common-ligatures"
PASS font-variant-numeric (type: discrete) has testAccumulation function
PASS font-variant-numeric: "oldstyle-nums" onto "lining-nums"
PASS font-variant-numeric: "lining-nums" onto "oldstyle-nums"
PASS font-variation-settings (type: fontVariationSettings) has testAccumulation function
FAIL font-variation-settings with composite type accumulate assert_equals: The value should be "wght" 2.2 at 250ms expected "\"wght\" 2.2" but got "\"wght\" 1.2"
PASS font-variation-settings (type: discrete) has testAccumulation function
PASS font-variation-settings: ""wdth" 5" onto ""wght" 1.1, "wdth" 1"
PASS font-variation-settings: ""wght" 1.1, "wdth" 1" onto ""wdth" 5"
PASS font-variation-settings: "normal" onto ""wdth" 5"
PASS font-variation-settings: ""wdth" 5" onto "normal"
PASS grid-auto-columns (type: discrete) has testAccumulation function
PASS grid-auto-columns: "5px" onto "1px"
PASS grid-auto-columns: "1px" onto "5px"
PASS grid-auto-flow (type: discrete) has testAccumulation function
PASS grid-auto-flow: "column" onto "row"
PASS grid-auto-flow: "row" onto "column"
PASS grid-auto-rows (type: discrete) has testAccumulation function
PASS grid-auto-rows: "5px" onto "1px"
PASS grid-auto-rows: "1px" onto "5px"
PASS grid-column-end (type: discrete) has testAccumulation function
PASS grid-column-end: "5" onto "1"
PASS grid-column-end: "1" onto "5"
PASS grid-column-start (type: discrete) has testAccumulation function
PASS grid-column-start: "5" onto "1"
PASS grid-column-start: "1" onto "5"
PASS grid-row-end (type: discrete) has testAccumulation function
PASS grid-row-end: "5" onto "1"
PASS grid-row-end: "1" onto "5"
PASS grid-row-start (type: discrete) has testAccumulation function
PASS grid-row-start: "5" onto "1"
PASS grid-row-start: "1" onto "5"
PASS grid-template-areas (type: discrete) has testAccumulation function
PASS grid-template-areas: "none" onto "". . a b" ". .a b""
FAIL grid-template-areas: "". . a b" ". .a b"" onto "none" assert_equals: The value should be ". . a b" ". .a b" at 0ms expected "\". . a b\" \". .a b\"" but got "\". . a b\" \". . a b\""
PASS grid-template-columns (type: discrete) has testAccumulation function
PASS grid-template-columns: "5px" onto "1px"
PASS grid-template-columns: "1px" onto "5px"
PASS grid-template-rows (type: discrete) has testAccumulation function
PASS grid-template-rows: "5px" onto "1px"
PASS grid-template-rows: "1px" onto "5px"
PASS hyphens (type: discrete) has testAccumulation function
PASS hyphens: "none" onto "manual"
PASS hyphens: "manual" onto "none"
PASS image-orientation (type: discrete) has testAccumulation function
PASS image-orientation: "from-image" onto "none"
FAIL image-orientation: "none" onto "from-image" assert_equals: The value should be none at 0ms expected "none" but got "from-image"
PASS isolation (type: discrete) has testAccumulation function
PASS isolation: "isolate" onto "auto"
PASS isolation: "auto" onto "isolate"
PASS justify-content (type: discrete) has testAccumulation function
PASS justify-content: "end" onto "start"
PASS justify-content: "start" onto "end"
PASS justify-items (type: discrete) has testAccumulation function
PASS justify-items: "end" onto "start"
PASS justify-items: "start" onto "end"
PASS justify-self (type: discrete) has testAccumulation function
PASS justify-self: "end" onto "start"
PASS justify-self: "start" onto "end"
PASS letter-spacing (type: length) has testAccumulation function
FAIL letter-spacing: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL letter-spacing: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS lighting-color (type: color) has testAccumulation function
FAIL lighting-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL lighting-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL lighting-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL lighting-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL lighting-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL lighting-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS list-style-image (type: discrete) has testAccumulation function
PASS list-style-image: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS list-style-image: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS list-style-position (type: discrete) has testAccumulation function
PASS list-style-position: "outside" onto "inside"
PASS list-style-position: "inside" onto "outside"
PASS list-style-type (type: discrete) has testAccumulation function
PASS list-style-type: "square" onto "circle"
PASS list-style-type: "circle" onto "square"
PASS marker-end (type: discrete) has testAccumulation function
PASS marker-end: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS marker-end: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS marker-mid (type: discrete) has testAccumulation function
PASS marker-mid: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS marker-mid: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS marker-start (type: discrete) has testAccumulation function
PASS marker-start: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS marker-start: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS mask (type: discrete) has testAccumulation function
PASS mask: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS mask: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS mask-type (type: discrete) has testAccumulation function
PASS mask-type: "luminance" onto "alpha"
PASS mask-type: "alpha" onto "luminance"
PASS mix-blend-mode (type: discrete) has testAccumulation function
PASS mix-blend-mode: "screen" onto "multiply"
PASS mix-blend-mode: "multiply" onto "screen"
PASS object-fit (type: discrete) has testAccumulation function
PASS object-fit: "contain" onto "fill"
PASS object-fit: "fill" onto "contain"
PASS order (type: integer) has testAccumulation function
FAIL order: integer assert_equals: The value should be -3 at 0ms expected "-3" but got "-2"
PASS outline-color (type: color) has testAccumulation function
FAIL outline-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL outline-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL outline-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL outline-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL outline-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL outline-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS outline-offset (type: length) has testAccumulation function
FAIL outline-offset: length assert_equals: The value should be 20px at 0ms expected "20px" but got "0px"
FAIL outline-offset: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "0px"
PASS outline-style (type: discrete) has testAccumulation function
PASS outline-style: "dotted" onto "none"
PASS outline-style: "none" onto "dotted"
PASS outline-width (type: length) has testAccumulation function
FAIL outline-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL outline-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS overflow-wrap (type: discrete) has testAccumulation function
PASS overflow-wrap: "break-word" onto "normal"
PASS overflow-wrap: "normal" onto "break-word"
PASS overflow-x (type: discrete) has testAccumulation function
PASS overflow-x: "hidden" onto "visible"
PASS overflow-x: "visible" onto "hidden"
PASS overflow-y (type: discrete) has testAccumulation function
PASS overflow-y: "hidden" onto "visible"
PASS overflow-y: "visible" onto "hidden"
PASS page-break-after (type: discrete) has testAccumulation function
PASS page-break-after: "auto" onto "always"
PASS page-break-after: "always" onto "auto"
PASS page-break-before (type: discrete) has testAccumulation function
PASS page-break-before: "auto" onto "always"
PASS page-break-before: "always" onto "auto"
PASS page-break-inside (type: discrete) has testAccumulation function
PASS page-break-inside: "avoid" onto "auto"
PASS page-break-inside: "auto" onto "avoid"
PASS paint-order (type: discrete) has testAccumulation function
FAIL paint-order: "stroke" onto "fill" assert_equals: The value should be stroke at 0ms expected "stroke" but got "stroke fill markers"
FAIL paint-order: "fill" onto "stroke" assert_equals: The value should be fill at 0ms expected "fill" but got "fill stroke markers"
PASS perspective (type: length) has testAccumulation function
FAIL perspective: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL perspective: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS perspective-origin (type: position) has testAccumulation function
PASS perspective-origin: length pair
PASS perspective-origin: length pair of rem
FAIL perspective-origin: position of percentage assert_equals: The value should be 1019.19px 0px at 0ms expected "1019.19px 0px" but got "548.797px 0px"
PASS pointer-events (type: discrete) has testAccumulation function
PASS pointer-events: "none" onto "fill"
PASS pointer-events: "fill" onto "none"
PASS position (type: discrete) has testAccumulation function
PASS position: "fixed" onto "absolute"
PASS position: "absolute" onto "fixed"
PASS quotes (type: discrete) has testAccumulation function
PASS quotes: ""‘" "’" "“" "”"" onto ""“" "”" "‘" "’""
PASS quotes: ""“" "”" "‘" "’"" onto ""‘" "’" "“" "”""
PASS resize (type: discrete) has testAccumulation function
PASS resize: "horizontal" onto "both"
PASS resize: "both" onto "horizontal"
PASS scroll-behavior (type: discrete) has testAccumulation function
PASS scroll-behavior: "smooth" onto "auto"
PASS scroll-behavior: "auto" onto "smooth"
PASS shape-outside (type: discrete) has testAccumulation function
PASS shape-outside: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS shape-outside: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS shape-rendering (type: discrete) has testAccumulation function
PASS shape-rendering: "crispEdges" onto "optimizeSpeed"
PASS shape-rendering: "optimizeSpeed" onto "crispEdges"
PASS stop-color (type: color) has testAccumulation function
FAIL stop-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL stop-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL stop-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL stop-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL stop-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL stop-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS stop-opacity (type: opacity) has testAccumulation function
FAIL stop-opacity: [0, 1] number assert_equals: The value should be 0.6 at 0ms expected "0.6" but got "0.3"
FAIL stop-opacity: [0, 1] number (clamped) assert_equals: The value should be 1 at 0ms expected "1" but got "0.3"
PASS stroke-dasharray (type: dasharray) has testAccumulation function
FAIL stroke-dasharray: dasharray assert_equals: The value should be 1, 2, 3, 4, 5 at 0ms expected "1, 2, 3, 4, 5" but got "1px, 2px, 3px, 4px, 5px"
PASS stroke-dasharray (type: discrete) has testAccumulation function
FAIL stroke-dasharray: "10, 20" onto "none" assert_equals: The value should be 10, 20 at 0ms expected "10, 20" but got "10px, 20px"
PASS stroke-dasharray: "none" onto "10, 20"
PASS stroke-linecap (type: discrete) has testAccumulation function
PASS stroke-linecap: "square" onto "round"
PASS stroke-linecap: "round" onto "square"
PASS stroke-linejoin (type: discrete) has testAccumulation function
PASS stroke-linejoin: "miter" onto "round"
PASS stroke-linejoin: "round" onto "miter"
PASS stroke-miterlimit (type: positiveNumber) has testAccumulation function
FAIL stroke-miterlimit: positive number assert_equals: The value should be 2.2 at 0ms expected "2.2" but got "1.1"
PASS stroke-opacity (type: opacity) has testAccumulation function
FAIL stroke-opacity: [0, 1] number assert_equals: The value should be 0.6 at 0ms expected "0.6" but got "0.3"
FAIL stroke-opacity: [0, 1] number (clamped) assert_equals: The value should be 1 at 0ms expected "1" but got "0.3"
PASS table-layout (type: discrete) has testAccumulation function
PASS table-layout: "fixed" onto "auto"
PASS table-layout: "auto" onto "fixed"
PASS text-align (type: discrete) has testAccumulation function
PASS text-align: "end" onto "start"
PASS text-align: "start" onto "end"
PASS text-align-last (type: discrete) has testAccumulation function
PASS text-align-last: "end" onto "start"
PASS text-align-last: "start" onto "end"
PASS text-anchor (type: discrete) has testAccumulation function
PASS text-anchor: "end" onto "middle"
PASS text-anchor: "middle" onto "end"
PASS text-combine-upright (type: discrete) has testAccumulation function
PASS text-combine-upright: "none" onto "all"
PASS text-combine-upright: "all" onto "none"
PASS text-decoration-color (type: color) has testAccumulation function
FAIL text-decoration-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL text-decoration-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL text-decoration-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL text-decoration-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL text-decoration-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL text-decoration-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS text-decoration-line (type: discrete) has testAccumulation function
PASS text-decoration-line: "overline" onto "underline"
PASS text-decoration-line: "underline" onto "overline"
PASS text-decoration-style (type: discrete) has testAccumulation function
PASS text-decoration-style: "dotted" onto "solid"
PASS text-decoration-style: "solid" onto "dotted"
PASS text-overflow (type: discrete) has testAccumulation function
PASS text-overflow: "ellipsis" onto "clip"
PASS text-overflow: "clip" onto "ellipsis"
PASS text-rendering (type: discrete) has testAccumulation function
PASS text-rendering: "optimizeLegibility" onto "optimizeSpeed"
PASS text-rendering: "optimizeSpeed" onto "optimizeLegibility"
PASS text-shadow (type: textShadowList) has testAccumulation function
FAIL text-shadow: shadow assert_equals: The value should be rgb(240, 240, 240) 20px 20px 20px at 0ms expected "rgb(240, 240, 240) 20px 20px 20px" but got "rgb(120, 120, 120) 10px 10px 10px"
PASS text-transform (type: discrete) has testAccumulation function
PASS text-transform: "uppercase" onto "capitalize"
PASS text-transform: "capitalize" onto "uppercase"
PASS touch-action (type: discrete) has testAccumulation function
PASS touch-action: "none" onto "auto"
PASS touch-action: "auto" onto "none"
PASS transform (type: transformList) has testAccumulation function
FAIL transform: translate assert_approx_equals: expected matrix(1,0,0,1,-100,0) but got matrix(1, 0, 0, 1, -200, 0): The value should be matrix(1,0,0,1,-100,0) at 0ms but got matrix(1, 0, 0, 1, -200, 0) expected -100 +/- 0.0001 but got -200
FAIL transform: rotate assert_approx_equals: expected matrix(0.7071067811865476,-0.7071067811865475,0.7071067811865475,0.7071067811865476,0,0) but got matrix(6.12323e-17, -1, 1, 6.12323e-17, 0, 0): The value should be matrix(0.7071067811865476,-0.7071067811865475,0.7071067811865475,0.7071067811865476,0,0) at 0ms but got matrix(6.12323e-17, -1, 1, 6.12323e-17, 0, 0) expected 0.7071067811865476 +/- 0.0001 but got 6.12323e-17
FAIL transform: scale assert_approx_equals: expected matrix(-2,0,0,-2,0,0) but got matrix(-3, 0, 0, -3, 0, 0): The value should be matrix(-2,0,0,-2,0,0) at 0ms but got matrix(-3, 0, 0, -3, 0, 0) expected -2 +/- 0.0001 but got -3
FAIL transform: skew assert_approx_equals: expected matrix(1,0.5773502691896257,-0.36397023426620234,1,0,0) but got matrix(1, 0.36397, -0.57735, 1, 0, 0): The value should be matrix(1,0.5773502691896257,-0.36397023426620234,1,0,0) at 0ms but got matrix(1, 0.36397, -0.57735, 1, 0, 0) expected 0.5773502691896257 +/- 0.0001 but got 0.36397
FAIL transform: rotate on translate assert_approx_equals: expected matrix(0,1,-1,0,100,0) but got matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0): The value should be matrix(0,1,-1,0,100,0) at 0ms but got matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0) expected 100 +/- 0.0001 but got 0
FAIL transform: translate on rotate assert_approx_equals: expected matrix(0,1,-1,0,100,0) but got matrix(1, 0, 0, 1, 100, 0): The value should be matrix(0,1,-1,0,100,0) at 0ms but got matrix(1, 0, 0, 1, 100, 0) expected 0 +/- 0.0001 but got 1
FAIL transform: rotate and translate on rotate assert_approx_equals: expected matrix(6.123233995736766e-17,1,-1,6.123233995736766e-17,0,0) but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0): The value should be matrix(6.123233995736766e-17,1,-1,6.123233995736766e-17,0,0) at 0ms but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0) expected 6.123233995736766e-17 +/- 0.0001 but got 0.707107
FAIL transform: rotate on rotate and translate assert_approx_equals: expected matrix(6.123233995736766e-17,1,-1,6.123233995736766e-17,6.123233995736766e-15,100) but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0): The value should be matrix(6.123233995736766e-17,1,-1,6.123233995736766e-17,6.123233995736766e-15,100) at 0ms but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0) expected 6.123233995736766e-17 +/- 0.0001 but got 0.707107
FAIL transform: matrix assert_approx_equals: expected matrix(0,1,-1,0,100,0) but got matrix(1, 0, 0, 1, 100, 0): The value should be matrix(0,1,-1,0,100,0) at 0ms but got matrix(1, 0, 0, 1, 100, 0) expected 0 +/- 0.0001 but got 1
FAIL transform: rotate3d assert_approx_equals: expected matrix3d(0.8535533905932737,0.1464466094067262,0.5,0,0.1464466094067262,0.8535533905932737,-0.5,0,-0.5,0.5,0.7071067811865476,0,0,0,0,1) but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 6.12323e-17, 0, 0, 0, 0, 1): The value should be matrix3d(0.8535533905932737,0.1464466094067262,0.5,0,0.1464466094067262,0.8535533905932737,-0.5,0,-0.5,0.5,0.7071067811865476,0,0,0,0,1) at 0ms but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 6.12323e-17, 0, 0, 0, 0, 1) expected 0.8535533905932737 +/- 0.0001 but got 0.5
FAIL transform: matrix3d assert_approx_equals: expected matrix3d(0.8535533905932737,0.1464466094067262,0.5,0,0.1464466094067262,0.8535533905932737,-0.5,0,-0.5,0.5,0.7071067811865476,0,0,0,0,1) but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 4.44089e-16, 0, 0, 0, 0, 1): The value should be matrix3d(0.8535533905932737,0.1464466094067262,0.5,0,0.1464466094067262,0.8535533905932737,-0.5,0,-0.5,0.5,0.7071067811865476,0,0,0,0,1) at 0ms but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 4.44089e-16, 0, 0, 0, 0, 1) expected 0.8535533905932737 +/- 0.0001 but got 0.5
FAIL transform: none assert_equals: dimension of the matrix: The value should be matrix3d(1,0,0,0,0,1,0,0,0,0,1,0,0,0,1,1) at 0ms but got matrix(1, 0, 0, 1, 0, 0) expected 16 but got 6
PASS transform: non-invertible matrices (non-invertible onto invertible)
PASS transform: non-invertible matrices (invertible onto non-invertible)
PASS transform: non-invertible matrices in matched transform lists (non-invertible onto invertible)
PASS transform: non-invertible matrices in matched transform lists (invertible onto non-invertible)
PASS transform: non-invertible matrices in mismatched transform lists (non-invertible onto invertible)
PASS transform: non-invertible matrices in mismatched transform lists (invertible onto non-invertible)
PASS transform-box (type: discrete) has testAccumulation function
FAIL transform-box: "border-box" onto "fill-box" assert_equals: The value should be border-box at 0ms expected "border-box" but got "fill-box"
PASS transform-box: "fill-box" onto "border-box"
PASS transform-style (type: discrete) has testAccumulation function
PASS transform-style: "preserve-3d" onto "flat"
PASS transform-style: "flat" onto "preserve-3d"
PASS rotate (type: rotateList) has testAccumulation function
FAIL rotate without rotation axes assert_equals: The value should be -45deg at 0ms expected "-45deg" but got "-90deg"
FAIL rotate with underlying transform assert_equals: The value should be 1 0 0 45deg at 0ms expected "1 0 0 45deg" but got "1 0 0 90deg"
FAIL rotate with different rotation axes assert_approx_equals: expected 0.707107 0 0.707107 90deg but got 1 0 1 90deg: The value should be 0.707107 0 0.707107 90deg at 500ms but got 1 0 1 90deg expected 0.707107 +/- 0.0001 but got 1
PASS translate (type: translateList) has testAccumulation function
FAIL translate assert_equals: The value should be -100px at 0ms expected "-100px" but got "-200px"
FAIL translate with transform assert_equals: The value should be -100px at 0ms expected "-100px" but got "-200px"
PASS scale (type: scaleList) has testAccumulation function
FAIL scale with two unspecified values assert_equals: The value should be -2 at 0ms expected "-2" but got "-3 1"
FAIL scale with one unspecified value assert_equals: The value should be -2 -2 at 0ms expected "-2 -2" but got "-3 -3"
FAIL scale assert_equals: The value should be 0 -1 -2 at 0ms expected "0 -1 -2" but got "-1 -2 -3"
PASS vector-effect (type: discrete) has testAccumulation function
PASS vector-effect: "non-scaling-stroke" onto "none"
PASS vector-effect: "none" onto "non-scaling-stroke"
PASS visibility (type: visibility) has testAccumulation function
FAIL visibility: onto "visible" assert_equals: The value should be visible at 1000ms expected "visible" but got "hidden"
PASS visibility: onto "hidden"
PASS white-space (type: discrete) has testAccumulation function
PASS white-space: "nowrap" onto "pre"
PASS white-space: "pre" onto "nowrap"
PASS word-break (type: discrete) has testAccumulation function
PASS word-break: "break-all" onto "keep-all"
PASS word-break: "keep-all" onto "break-all"
PASS word-spacing (type: lengthPercentageOrCalc) has testAccumulation function
PASS word-spacing: length
PASS word-spacing: length of rem
FAIL word-spacing: percentage assert_equals: The value should be 130% at 0ms expected "130%" but got "0px"
FAIL word-spacing: units "%" onto "px" assert_equals: The value should be calc(10px + 10%) at 0ms expected "calc(10px + 10%)" but got "10px"
FAIL word-spacing: units "px" onto "%" assert_equals: The value should be calc(10px + 10%) at 0ms expected "calc(10px + 10%)" but got "10px"
FAIL word-spacing: units "rem" onto "%" assert_equals: The value should be calc(20px + 10%) at 0ms expected "calc(20px + 10%)" but got "20px"
FAIL word-spacing: units "%" onto "rem" assert_equals: The value should be calc(20px + 10%) at 0ms expected "calc(20px + 10%)" but got "20px"
FAIL word-spacing: units "rem" onto "em" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL word-spacing: units "em" onto "rem" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL word-spacing: units "calc" onto "px" assert_equals: The value should be calc(30px + 20%) at 0ms expected "calc(30px + 20%)" but got "10px"
FAIL word-spacing: calc assert_equals: The value should be calc(30px + 30%) at 0ms expected "calc(30px + 30%)" but got "0px"
Harness: the test ran to completion.
